index.mjs splitted into internal.mjs and global.mjs
