+++ global.mjs
@@ -1,0 +1,3 @@
+import {populateGlobal} from './internal.mjs'
+
+populateGlobal()
